demo/requirements.txt
Pillow 
numpy 
decorator 
scipy 
attrs 
pytest
